requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.11.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.1
neo4j==5.17.0
PyJWT==2.8.0
werkzeug==3.0.1
tenacity==8.2.3
openai==1.12.0
flask==2.3.3
email-validator==2.2.0 
requests==2.31.0
pytest==7.4.3 